replace old simulation with new, add requirements.txt, re-activate test target in makefile, add test to travis ci build
